requirements.txt
viam-sdk
ouster-sdk
open3d